requirements.txt
# ============================================================
# FedGreen-C 项目依赖
# ============================================================

# 基础科学计算
numpy==2.4.3
pandas==2.3.3
scipy==1.17.1
scikit-learn==1.8.0
statsmodels==0.14.4

# 深度学习
torch==2.10.0
torchvision==0.25.0
torchaudio==2.10.0
torch-geometric==2.7.0
torch_cluster==1.6.3+pt22cpu
torch_scatter==2.1.2+pt22cpu
torch_sparse==0.6.18+pt22cpu
torch_spline_conv==1.2.2+pt22cpu

# 机器学习
xgboost==3.2.0
joblib==1.5.3
threadpoolctl==3.6.0

# 可视化
matplotlib==3.10.8
seaborn==0.13.2
plotly==6.6.0
altair==6.0.0
folium==0.20.0
streamlit==1.55.0
streamlit-folium==0.26.2
pydeck==0.9.1

# 数据处理
openpyxl==3.1.5
pyarrow==23.0.1
pillow==12.1.1
xxhash==3.6.0

# 网络请求
requests==2.32.5
urllib3==2.6.3
certifi==2026.2.25
idna==3.11
charset-normalizer==3.4.6

# 时间处理
python-dateutil==2.9.0.post0
pytz==2026.1.post1
tzdata==2025.3

# 节假日
holidays==0.55

# 优化和调参
optuna==4.8.0
sqlalchemy==2.0.48
alembic==1.18.4
Mako==1.3.10
greenlet==3.3.2
colorlog==6.10.1

# 文件处理
toml==0.10.2
pyyaml==6.0.2
pickle-mixin==1.0.2

# 系统监控
psutil==7.2.2
watchdog==6.0.0
tqdm==4.67.3

# Web 框架
Jinja2==3.1.6
MarkupSafe==3.0.3
tornado==6.5.5
blinker==1.9.0
click==8.3.1
colorama==0.4.6

# Git 相关
gitdb==4.0.12
GitPython==3.1.46
smmap==5.0.3

# 其他
attrs==25.4.0
branca==0.8.2
cachetools==7.0.5
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
mpmath==1.3.0
multidict==6.7.1
narwhals==2.18.0
networkx==3.6.1
packaging==26.0
propcache==0.4.1
protobuf==6.33.5
pyparsing==3.3.2
referencing==0.37.0
rpds-py==0.30.0
setuptools==82.0.1
six==1.17.0
sympy==1.14.0
tenacity==9.1.4
typing_extensions==4.15.0
xyzservices==2025.11.0
yarl==1.23.0

# 可解释性分析
shap==0.51.0
slicer==0.0.8
numba==0.64.0
cloudpickle==3.1.2
llvmlite==0.46.0

# TensorFlow (用于 SHAP DeepExplainer)
tensorflow-cpu==2.21.0
grpcio==1.78.0
h5py==3.14.0
keras==3.13.2
ml_dtypes==0.5.4
opt_einsum==3.4.0
gast==0.7.0
libclang==18.1.1
wrapt==2.1.2
termcolor==3.3.0
